\name{rD3plot-package}
\alias{rD3plot-package}
\docType{package}
\title{The rD3plot package.}
\description{
Creates interactive analytic graphs with 'R'. It joins the data analysis power of R and the visualization libraries of JavaScript in one package. The package provides interactive networks, timelines, barplots, image galleries and evolving networks. Graphs are represented as D3 graphs embedded in a web page ready for its interactive analysis and exploration
}
