\name{shiny_rd3}
\alias{shiny_rd3}
\title{Include rD3plot graphs in Shiny.}
\description{Load a rD3plot graph to display in Shiny.}
\usage{
shiny_rd3(x)
}
\arguments{
  \item{x}{is a \code{network_rd3}, \code{barplot_rd3} or \code{timeplot_rd3} object.}
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. 
}
